splitted profiles msg_server
